+++ src/libs/shared/src/ic/mod.rs
@@ -1,5 +1,5 @@
 pub mod api;
 pub mod call;
-mod utils;
+mod trap;
 
-pub use utils::*;
+pub use trap::*;
